Created loading page and moved login from main to login
